requirements.txt
grpcio
grpcio-reflection
protobuf
numpy
opencv-python-headless